Set available addresses when Azure returns nil IP configs
In some scenarios we noticed that in describe VPCs API call, azure can
return nill IP configs for some subnets, even when there are IP configs
allocated.

This results in the operator initializing subnets with no available
addresses. When this happens operator falls back to selecting other
subnets to allocate IPs from, which is also not supported in Azure and
results in VMScaleSetIpConfigurationsOnSameNicCannotUseDifferentSubnets
error code. Instead, when IP config is nil, initialize subnets with
total IP capacity.

Signed-off-by: Hemanth Malla <hemanth.malla@datadoghq.com>