wireguard: reduce syscall ConfigureDevice when not needed
Whenever we call the `updatePeerByConfig` in the Wireguard agent, we
always perform `ConfigureDevice` even in case there are no changes to
introduce meaning:

1. no new allowedIPs to add;
2. no new endpoint;
3. no new key.

This could result in many unneeded syscalls being used anyway, such as when:

* finishing restoring IPs for a peer with no IP to be removed;
* handling an IPCache delete event with no new IP to add.

In `UpserPeer()` we want to always configure the device, since at least one
of the three events above (1-3) is met. On the other hand, when restoring
allowedIPs or when handling an IPCache event, we can avoid configuring the
device when no relevant changes are recorded. Note that IPs that needs to
be removed would be correctly handled and removed in any case.

Signed-off-by: Simone Magnani <simone.magnani@isovalent.com>