bpf: wireguard: IS_BPF_WIREGUARD in bpf_host for tracing point
This commit modifies the tracing observation point in cil_from_netdev
in case the program is injected into the Wireguard device. We do that
to attach `cil_from_netdev` to the ingress hook of the wg device given
a set of conditions. In that case, the tracing system would still use
TRACE_FROM_NETWORK as observation point: let's update it with the new
TRACE_FROM_CRYPTO one (a from-network even would already be traced from
the native device).

Note: we do not recompile the program when attaching it into the wg
interface. In this case, IS_BPF_WIREGUARD is a config variable set to
false by default, which the loader will update to true (1) when detecting
that the interface to attach the program to is wg.

Signed-off-by: Simone Magnani <simone.magnani@isovalent.com>