bpf: overlay: retain security identity when forwarding service replies
Consider a configuration with
* overlay routing mode
* no SocketLB (so that pod-to-remote-nodeport is translated at the
  remote node)
* HostFW enabled (so that pod-to-remote-node traffic flows via the
  overlay)

When a pod accesses a nodeport service on a remote node, in such a config
both the forward *and* reply direction will flow via the overlay network.
The LB node handles the SVC translation (and reverse translation) inside
the `from-overlay` program.
* the forward path is handled by tail_nodeport_nat_egress_ipv*(). It
  always inserts WORLD_ID as security identity, even for the case of such
  an in-cluster connection.
* the reply path is handled by nodeport_rev_dnat_ingress_ipv*(). It
  currently also inserts WORLD_ID as security identity.

Considering the reply path - when `from_overlay` handles the packet, it
still has access to the backend's actual security identity (as this was
transferred in the tunnel header). Let's forward this identity all the
way to the client, instead of reporting a vague WORLD_ID.

As we're dealing with reply traffic, the network policy code on the
client side won't see an actual difference.

Alternatively we could also report REMOTE_NODE_ID here, which should be
symmetric with the forward path (as it accesses a nodeport on a remote
node).

Signed-off-by: Julian Wiedmann <jwi@isovalent.com>